qt6升级

+++ Code/Tools/LuaIDE/Source/LUA/RecentFiles.cpp
@@ -45,7 +45,7 @@ void WriteRecentFiles(const QStringList& recentFiles)
 
     settings.beginGroup(LUAEDITOR_GROUPNAME);
     settings.beginWriteArray(LUAEDITOR_SETTINGS_RECENT_FILES_KEY);
-    int count = std::min(recentFiles.size(),
+    int count = std::min(static_cast<int>(recentFiles.size()),
             LUAEDITOR_SETTINGS_RECENT_FILES_COUNT_MAX);
 
     // QSettings -> QStringList.
