+++ Code/Tools/ProjectManager/Source/CreateAGemScreen.cpp
@@ -285,7 +285,7 @@ namespace O3DE::ProjectManager
         QStringList platformOptions;
         
         //input the platform list in reverse alphabetical order
-        for(int i = GemInfo::NumPlatforms-1; i >= 0; i--)
+        for(int i = static_cast<int>(GemInfo::NumPlatforms)-1; i >= 0; i--)
         {
             const GemInfo::Platform platform = static_cast<GemInfo::Platform>(1 << i);
             if(platform & m_platformSupportMask)
